+++ container-engine-lib/docs/changelog.md
@@ -1,4 +1,10 @@
 # TBD
+### Features
+* Added `ExecResult` object to represent the result of commands execution inside an instance
+
+### Breaking Changes
+* Replaced the first returned var type `map[service.ServiceGUID]bool` in `RunUserServiceExecCommands` and `RunNetworkingSidecarExecCommands` with `map[service.ServiceGUID]*exec_result.ExecResult`
+  * Users can get the new returned map and use the `ExecResult` object to obtain the execution exit code and output 
 
 # 0.10.4
 ### Features